+++ src/gui/mod.rs
@@ -26,9 +26,9 @@ use ajour_widgets::header;
 use async_std::sync::{Arc, Mutex};
 use chrono::{DateTime, NaiveDateTime, Utc};
 use iced::{
-    button, pick_list, scrollable, text_input, Align, Application, Button, Clipboard, Column,
-    Command, Container, Element, HorizontalAlignment, Length, PickList, Row, Scrollable, Settings,
-    Space, Subscription, Text, TextInput,
+    button, pick_list, scrollable, slider, text_input, Align, Application, Button, Clipboard,
+    Column, Command, Container, Element, HorizontalAlignment, Length, PickList, Row, Scrollable,
+    Settings, Space, Subscription, Text, TextInput,
 };
 use image::ImageFormat;
 use isahc::http::Uri;
@@ -41,16 +41,16 @@ use strfmt::strfmt;
 use element::{DEFAULT_FONT_SIZE, DEFAULT_PADDING};
 static WINDOW_ICON: &[u8] = include_bytes!("../../resources/windows/ajour.ico");
 
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug)]
 pub enum State {
-    Start,
     Ready,
     Loading,
+    Error(anyhow::Error),
 }
 
 impl Default for State {
     fn default() -> Self {
-        State::Start
+        State::Ready
     }
 }
 
@@ -62,6 +62,12 @@ pub enum Mode {
     Catalog,
     Settings,
     About,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum Confirm {
+    DeleteAddon,
+    DeleteSavedVariables,
 }
 
 impl std::fmt::Display for Mode {
@@ -84,7 +90,10 @@ impl std::fmt::Display for Mode {
 #[derive(Debug, Clone)]
 #[allow(clippy::large_enum_variant)]
 pub enum Interaction {
-    Delete(String),
+    DeleteAddon(),
+    ConfirmDeleteAddon(String),
+    DeleteSavedVariables(),
+    ConfirmDeleteSavedVariables(String),
     Expand(ExpandType),
     Ignore(String),
     SelectBackupDirectory(),
@@ -137,6 +146,7 @@ pub enum Interaction {
     ToggleStartClosedToTray(bool),
     ThemeUrlInput(String),
     ImportTheme,
+    CompressionLevelChanged(i32),
 }
 
 #[derive(Debug)]
@@ -235,12 +245,17 @@ pub struct Ajour {
     addons_search_state: AddonsSearchState,
     wow_directories: Vec<WowDirectoryState>,
     default_backup_compression_format: pick_list::State<CompressionFormat>,
+    pending_confirmation: Option<Confirm>,
+    zstd_compression_level_slider_state: slider::State,
 }
 
 impl Default for Ajour {
     fn default() -> Self {
+        let mut state = HashMap::new();
+        state.insert(Mode::Catalog, State::Loading);
+
         Self {
-            state: [(Mode::Catalog, State::Loading)].iter().cloned().collect(),
+            state,
             error: None,
             mode: Mode::MyAddons(Flavor::Retail),
             addons: HashMap::new(),
@@ -297,6 +312,8 @@ impl Default for Ajour {
                 })
                 .collect::<Vec<WowDirectoryState>>(),
             default_backup_compression_format: Default::default(),
+            pending_confirmation: None,
+            zstd_compression_level_slider_state: Default::default(),
         }
     }
 }
@@ -533,6 +550,7 @@ impl Application for Ajour {
                         &self.config,
                         &column_config,
                         is_odd,
+                        &self.pending_confirmation,
                     );
 
                     // Adds the addon data cell to the scrollable.
@@ -970,6 +988,7 @@ impl Application for Ajour {
                     &mut self.theme_state,
                     &mut self.scale_state,
                     &mut self.backup_state,
+                    &mut self.zstd_compression_level_slider_state,
                     &mut self.default_backup_compression_format,
                     &mut self.column_settings,
                     &column_config,
@@ -1001,19 +1020,9 @@ impl Application for Ajour {
 
         let container: Option<Container<Message>> = match self.mode {
             Mode::MyAddons(flavor) => {
-                let state = self
-                    .state
-                    .get(&Mode::MyAddons(flavor))
-                    .cloned()
-                    .unwrap_or_default();
+                let state = self.state.get(&Mode::MyAddons(flavor));
                 match state {
-                    State::Start => Some(element::status::data_container(
-                        color_palette,
-                        &localized_string("setup-ajour-title")[..],
-                        &localized_string("setup-ajour-description")[..],
-                        Some(&mut self.onboarding_directory_btn_state),
-                    )),
-                    State::Loading => {
+                    Some(State::Loading) => {
                         let flavor = flavor.to_string().to_lowercase();
                         let mut vars = HashMap::new();
                         vars.insert("flavor".to_string(), &flavor);
@@ -1026,7 +1035,7 @@ impl Application for Ajour {
                             None,
                         ))
                     }
-                    State::Ready => {
+                    Some(State::Ready) => {
                         let flavor = flavor.to_string().to_lowercase();
                         let mut vars = HashMap::new();
                         vars.insert("flavor".to_string(), &flavor);
@@ -1043,26 +1052,38 @@ impl Application for Ajour {
                             None
                         }
                     }
+                    Some(State::Error(error)) => {
+                        let error_title = format!("{}", error);
+                        let mut error_description = String::new();
+                        let cause = error.chain().last();
+
+                        if let Some(cause) = cause {
+                            error_description.push_str(&format!("Caused by {}", cause)[..]);
+                        }
+
+                        Some(element::status::data_container(
+                            color_palette,
+                            &error_title,
+                            &error_description,
+                            None,
+                        ))
+                    }
+                    None => Some(element::status::data_container(
+                        color_palette,
+                        &localized_string("setup-ajour-title")[..],
+                        &localized_string("setup-ajour-description")[..],
+                        Some(&mut self.onboarding_directory_btn_state),
+                    )),
                 }
             }
             Mode::Settings => None,
             Mode::About => None,
             Mode::Install => None,
             Mode::MyWeakAuras(flavor) => {
-                let state = self
-                    .state
-                    .get(&Mode::MyWeakAuras(flavor))
-                    .cloned()
-                    .unwrap_or_default();
+                let state = self.state.get(&Mode::MyWeakAuras(flavor));
 
                 match state {
-                    State::Start => Some(element::status::data_container(
-                        color_palette,
-                        &localized_string("setup-weakauras-title")[..],
-                        &localized_string("setup-weakauras-description")[..],
-                        None,
-                    )),
-                    State::Loading => {
+                    Some(State::Loading) => {
                         let flavor = flavor.to_string().to_lowercase();
                         let mut vars = HashMap::new();
                         vars.insert("flavor".to_string(), &flavor);
@@ -1075,7 +1096,7 @@ impl Application for Ajour {
                             None,
                         ))
                     }
-                    State::Ready => {
+                    Some(State::Ready) => {
                         if !has_auras {
                             let flavor = flavor.to_string();
                             let mut vars = HashMap::new();
@@ -1092,19 +1113,56 @@ impl Application for Ajour {
                             None
                         }
                     }
+                    Some(State::Error(error)) => {
+                        let error_title = format!("{}", error);
+                        let mut error_description = String::new();
+                        let cause = error.chain().last();
+
+                        if let Some(cause) = cause {
+                            error_description.push_str(&format!("Caused by {}", cause)[..]);
+                        }
+
+                        Some(element::status::data_container(
+                            color_palette,
+                            &error_title,
+                            &error_description,
+                            None,
+                        ))
+                    }
+                    None => Some(element::status::data_container(
+                        color_palette,
+                        &localized_string("setup-weakauras-title")[..],
+                        &localized_string("setup-weakauras-description")[..],
+                        None,
+                    )),
                 }
             }
             Mode::Catalog => {
-                let state = self.state.get(&Mode::Catalog).cloned().unwrap_or_default();
+                let state = self.state.get(&Mode::Catalog);
                 match state {
-                    State::Start => None,
-                    State::Loading => Some(element::status::data_container(
+                    Some(State::Loading) => Some(element::status::data_container(
                         color_palette,
                         &localized_string("loading")[..],
                         &localized_string("loading-catalog")[..],
                         None,
                     )),
-                    State::Ready => None,
+                    Some(State::Error(error)) => {
+                        let error_title = format!("{}", error);
+                        let mut error_description = String::new();
+                        let cause = error.chain().last();
+
+                        if let Some(cause) = cause {
+                            error_description.push_str(&format!("Caused by {}", cause)[..]);
+                        }
+
+                        Some(element::status::data_container(
+                            color_palette,
+                            &error_title,
+                            &error_description,
+                            None,
+                        ))
+                    }
+                    _ => None,
                 }
             }
         };
@@ -1985,6 +2043,7 @@ pub enum BackupFolderKind {
     WTF,
     Config,
     Screenshots,
+    Fonts,
 }
 
 #[derive(Default)]